# server/swagger.yaml
openapi: 3.0.0
info:
  title: QRio Event Management API
  version: 1.0.0
  description: |
    This is the official API documentation for the QRio Event Management System.

    Features:
    - User Registration, Email Verification & Authentication
    - Event Listing & Admin Management
    - Blog Posting with Comments & Likes
    - Event Registration with QR Ticketing System
servers:
  - url: http://localhost:5000
    description: Local Development Server
...
paths:
...
  /admin/events/add:
    post:
      tags: [Event]
      summary: Create a new event (admin)
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                date:
                  type: string
                  format: date
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Event created

  # ------------------ BLOG ROUTES ------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT